src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/text/reference/ArrayReferenceDoubleTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.array.text.ArrayTextParser;
import org.openqa.selenium.WebElement;

public class ArrayReferenceDoubleTextParser extends ArrayTextParser {

    public ArrayReferenceDoubleTextParser() {
        super(Double.class, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Double.parseDouble(element.getText());
    }

}