src/test/java/com/github/brunomndantas/jscrapper/support/parser/array/text/reference/ArrayReferenceDoubleTextParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class ArrayReferenceDoubleTextParserTest {

    @Test
    public void constructorTest() {
        ArrayReferenceDoubleTextParser parser = new ArrayReferenceDoubleTextParser();
        assertSame(Double.class, parser.getKlass());
        assertEquals(null, parser.getDefaultValue());
    }

    @Test
    public void parseElement() throws Exception {
        ArrayReferenceDoubleTextParser parser = new ArrayReferenceDoubleTextParser();

        assertEquals(2d, parser.parseElement(new DummyElement(){
            @Override
            public String getText() {
                return "2";
            }
        }));
    }

}